@@ -20,8 +20,10 @@
     "start": "run-p start:stencil start:storybook",
     "start:stencil": "stencil build --docs --dev --watch --serve --port 3334",
     "start:storybook": "start-storybook -p 6007 -s www",
-    "test": "NODE_ICU_DATA=node_modules/full-icu stencil test --spec",
-    "test:watch": "NODE_ICU_DATA=node_modules/full-icu stencil test --spec --watchAll",
+    "test": "NODE_ICU_DATA=node_modules/full-icu stencil test --spec -c stencil.testing-config.ts",
+    "test:e2e": "NODE_ICU_DATA=node_modules/full-icu stencil test --e2e -c stencil.testing-config.ts",
+    "test:watch": "npm run test -- --watchAll",
+    "test:e2e:watch": "npm run test:e2e -- --watchAll",
     "generate": "stencil generate",
     "build": "run-s build:stencil-ci build:storybook",
     "build:stencil-ci": "stencil build --ci -c stencil.build.config.ts",
@@ -34,40 +36,41 @@
     "i18next-http-backend": "^1.1.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.13",
-    "@stencil/core": "^2.4.0",
-    "@stencil/eslint-plugin": "0.3.1",
-    "@stencil/store": "1.3.0",
-    "@storybook/addon-a11y": "^6.1.17",
-    "@storybook/addon-actions": "^6.1.17",
+    "@babel/core": "^7.17.5",
+    "@stencil/core": "^2.14.0",
+    "@stencil/eslint-plugin": "0.4.0",
+    "@stencil/store": "1.5.0",
+    "@storybook/addon-a11y": "^6.4.19",
+    "@storybook/addon-actions": "^6.4.19",
     "@storybook/addon-console": "^1.2.3",
-    "@storybook/addon-essentials": "^6.1.17",
-    "@storybook/addon-links": "^6.1.17",
-    "@storybook/web-components": "^6.1.17",
+    "@storybook/addon-essentials": "^6.4.19",
+    "@storybook/addon-links": "^6.4.19",
+    "@storybook/web-components": "^6.4.19",
     "@types/events": "^3.0.0",
-    "@types/jest": "26.0.20",
-    "@types/lodash-es": "4.17.4",
-    "@types/puppeteer": "5.4.3",
-    "@typescript-eslint/eslint-plugin": "4.15.0",
-    "@typescript-eslint/parser": "4.15.0",
-    "babel-loader": "^8.2.2",
-    "eslint": "7.19.0",
-    "eslint-config-prettier": "7.2.0",
-    "eslint-plugin-prettier": "3.3.1",
-    "eslint-plugin-react": "7.22.0",
-    "full-icu": "1.3.1",
+    "@types/jest": "27.4.0",
+    "@types/lodash-es": "4.17.6",
+    "@typescript-eslint/eslint-plugin": "^4.33.0",
+    "@typescript-eslint/parser": "^4.33.0",
+    "babel-loader": "^8.2.3",
+    "chromium": "^3.0.3",
+    "eslint": "^7.32.0",
+    "eslint-config-prettier": "8.4.0",
+    "eslint-plugin-prettier": "4.0.0",
+    "eslint-plugin-react": "7.28.0",
+    "full-icu": "1.4.0",
     "i18next": "^19.8.7",
-    "jest": "26.6.3",
-    "jest-cli": "26.6.3",
+    "jest": "27.5.1",
+    "jest-cli": "27.5.1",
     "jest-date-mock": "1.0.8",
-    "lit-html": "^1.3.0",
-    "lodash-es": "4.17.20",
+    "lit-html": "^2.2.0",
+    "lodash-es": "4.17.21",
     "npm-run-all": "4.1.5",
-    "prettier": "2.2.1",
-    "puppeteer": "5.5.0",
+    "prettier": "2.5.1",
+    "puppeteer": "10.4.0",
+    "puppeteer-core": "10.4.0",
     "rimraf": "3.0.2",
-    "rxjs-marbles": "^6.0.1",
-    "typescript": "4.1.5"
+    "rxjs-marbles": "^7.0.1",
+    "typescript": "4.5.5"
   },
   "license": "MIT"
 }
